move Worker to its own file
